refactor modal directory placed under the ui directory
