+++ src/components/Archive/SoldItem/index.js
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
 import { Link, useParams } from 'react-router-dom'
 import moment from 'moment'
-import DeleteSoldItemModal from '../../Modals/DeleteSoldItemModal'
+import DeleteSoldItemModal from '../../ui/Modals/DeleteSoldItemModal'
 import Unauthorized from '../../ui/Alerts/Unauthorized'
 import Loading from '../../Spinners/Loading'
 import { SendGetRequest, SendPostRequest } from '../../../helpers/SendApiRequest'
