+++ src/components/FAQs/FAQs/index.js
@@ -2,7 +2,7 @@ import { React, useEffect, useState } from 'react'
 import { Table } from 'react-bootstrap'
 import { Link } from 'react-router-dom'
 import moment from 'moment'
-import DeleteGuideModal from '../../Modals/DeleteGuideModal'
+import DeleteGuideModal from '../../ui/Modals/DeleteGuideModal'
 import Loading from '../../Spinners/Loading'
 import classes from './index.module.css'
